Image upload now works
The client can now add images when creating a horse post. backend hosts images on Imgur and then sends the links to the front end which is added to horse data. The resulting data is then sent to the backend via the useAddHorses hook that uploads the horse to the database collection.